requirements.txt
Package          Version
---------------- -------
blinker          1.6.2
click            8.1.3
dominate         2.8.0
Flask            2.3.2
Flask-Bootstrap  3.3.7.1
Flask-Bootstrap4 4.0.2
Flask-MySQLdb    1.0.1
itsdangerous     2.1.2
Jinja2           3.1.2
MarkupSafe       2.1.3
mysqlclient      2.1.1
pip              23.1.2
PyYAML           6.0
setuptools       65.5.1
visitor          0.1.3
Werkzeug         2.3.6
wheel            0.38.4
